less codespeed exe variants
I now only maintain exe names like

  perl-5.13
  perl-5.14
  perl-5.15

instead of the last sub-version-number
in codespeed because otherwise there
is too much info fragmentation there.

One such exe currently covers one year
of development which is an ok focus of interest.

On 5.8 it will be too much in one exe but then
again I do not expect to cover this as exhaustive
as the current versions.