rework the cpu stabilization code
Now done via external script which we can call
via sudo.

The script is only called when run with --stabilize-cpu.

With this separation we can now more clearly address the
measurements on a current normal system and the continuous
tracking of the Perl speed over time on a dedicated fix system.