drop checking result receiver
The result reporting is now more light-weight
and just sends via HTTP, so we do not want to
use the heavy-weight BA::Lib anymore. And it also
was not a reliable check anyway because the result
store check was way too early.

Now we have a more robust result writing in case
reporting fails, so we can drop the fragile check.